+++ CHANGELOG.md
@@ -4,10 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - 2014-05-31
+## [0.5.2] - 2019-05-04
+### Added
+- Custom Grid System on X Axis
+- Custom Grid System on Z Axis
+- Add Different Size by Axis
+
+### Changed
+- Turn off the grid when cellSize = 0;
+
+## [0.1.0] - 2019-05-04
 ### Added
 - Custom Grid System on Y Axis
 - Snap to Grid option
 - Static Grid option
 
-[0.0.1]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.0.1
+[0.1.0]: https://github.com/DiazTeo/LevelGrid/tree/0.5.2
+[0.1.0]: https://github.com/DiazTeo/LevelGrid/tree/0.1.0